Experimentos/requirements.txt
shap~=0.40.0
numpy~=1.22.4
pandas~=1.4.2
seaborn~=0.11.2
matplotlib~=3.5.2
scikit-learn~=1.1.1
tensorflow~=2.9.1
scipy~=1.8.1
Pillow~=9.1.1